Delete song as host functionality working
play pause and skip are working. Pretty sure they are working on the
lock screen but that will need to be checked. When a song changes
during the parallax scroll the image gets fucked, this can probably be
fixed by setting the frame of the image or pinning the image to the top
of the view. Purple indicator for self posted songs needs to be re
implemented.